requirements.txt
python-telegram-bot==21.0.1
google-cloud-texttospeech==2.16.3
elevenlabs==1.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
